adjustments
adjustments to the trainer hill remaining unchanged normal mode trainers and some small tweaks to the depths, now both facilities should be fully functional in normal mode and just in need of balance tweaks (and finishing the expert modes). also made the battle pass obtainable so that players can actually challenge these facilities in a playthrough.