split path sailing
starting work on allowing the player to control the boat when travelling bewteen petalburg and slateport. this means the player will have to fight the various swimmers, so that taking the mauville path doesn't result in way more trainers than the dewford path. also its just cool. currently only going from petalburg to dewford is done.